Move /lib/fetch-teams to /functions/fetch-teams
