+++ src/index.ts
@@ -1,3 +1,3 @@
-import { fetchTeams } from './lib/fetch-teams.js';
+import { fetchTeams } from './functions/fetch-teams.js';
 
 await fetchTeams();
